refactor: move game logic into custom hook
